Data table schema object

An object that describes the schema of a data table, like a worksheet's summary data or underlying data, or a logical table from a data source.

Fields

id

Character vector indicating the ID of the underlying data table or logical table. (Not present for summary data.)

caption

Character vector with a human-readable description of the underlying data table or logical table. (Not present for summary data.)

name

Character vector with either "Underlying Data Table" or "Summary Data Table" (yes, literally those strings; not sure why this field is called name but it comes from the Tableau Extension API).

columns

A data frame that describes the columns in this table. Each data frame row describes a column in the data table. The data frame contains these columns:

dataType - character - "bool", "date", "date-time", "float", "int", "spatial", or "string".

fieldName - character - The name of the column.

index - integer - The column number.

isReferenced - logical - If TRUE, then the column is referenced in the worksheet.

See also