Specify options for tableau_extension

Use this function to modify the configuration options available for your Tableau extension. Pass the resulting value as the options argument to tableau_extension().

ext_options(
  config_width = 600,
  config_height = 400,
  prompt_for_config = TRUE,
  use_theme = TRUE,
  ...
)

Arguments

config_width, config_height
Numeric value specifying the initial width and height of the extension's configuration dialog (if any). In pixels.

prompt_for_config
If the extension provides a configuration dialog (i.e. tableau_extension() is called with a config_ui argument), prompt_for_config=TRUE (the default) means that a newly added extension should not even attempt to render its ui and server in the dashboard; instead, a message instructing the Tableau user to use the configuration dialog is displayed instead.

Use prompt_for_config=FALSE if the extension is able to run even without initial configuration.

use_theme
By default, shinytableau applies a customized version of Bootstrap 4 CSS to your extension's UI and config UI; the customizations are intended to complement Tableau's own UI conventions (though we cannot replicate them exactly due to licensing issues). Pass FALSE if you want to omit shinytableau's styles and just use your own.

...
Options to pass through to shiny::runApp (e.g. port, launch.browser, host, quiet). For local development purposes, it's a good idea to assign port to a hardcoded number between 1025 and 49151 that is unique for each extension.

Value

A list object, suitable for passing to tableau_extension(options = ...).