Create a hyperlink to embedding instructions

Use this function to create a link that standalone = TRUE apps can present in their UI, that will bring the user to the Tableau extension information page (where the manifest information and .trex file download button are found).

tableau_install_link(
  ...,
  target = "_blank",
  button = c("no", "primary", "default")
)

Arguments

...
Attributes and children of the element; these are passed through to htmltools::a().

target
The value of the target attribute. The default value of "_blank" will cause the link to open in a new browser tab or window. Pass NULL or "_self" instead to open in the current frame, or "_top" to use the current browser tab but to break out of any iframe.

button
If "primary" or "default", Bootstrap CSS classes will be added to make the link look like a button.

Value

An HTML object, suitable for including in Shiny UI.

See also

Examples

ui <- function(req) {
  fluidPage(

    # If we're not currently running in a Tableau dashboard, insert a link
    if (!tableau_is_embedded()) {
      absolutePanel(top = 10, right = 10,
        tableau_install_link("Embed this app in Tableau!")
      )
    },

    # Other UI...
  )
}