Get info about available worksheets in this Tableau dashboard

For advanced use only; most shinytableau extensions should use the choose_data() module to allow the user to specify a worksheet.

tableau_worksheets(session = shiny::getDefaultReactiveDomain())

tableau_worksheet_info(name, session = shiny::getDefaultReactiveDomain())

Arguments

session
The Shiny session object. (You should probably just use the default.)

name
A worksheet name, as returned by tableau_worksheets().

Value

tableau_worksheets() returns a character vector whose elements are worksheet names. Note that only worksheets that are included on the same dashboard will be listed, and these are the only worksheets we can access.

tableau_worksheet_info() returns metadata for a specific worksheet. The return value is a named list that contains the following fields:

name - The name of the worksheet.

summary - The data table schema object for the worksheet's summary-level data table.

dataSourceIds - Character vector of data source IDs used by this worksheet. See tableau_datasource_info().

underlyingTables - Unnamed list, each element is a data table schema object of one of the worksheet's underlying data tables.