Create a skeleton YAML manifest file

Tableau extensions require a manifest and the use of a YAML manifest file in shinytableau allows for a separation of concerns when developing an extension. With the yaml_skeleton() function, we can write a YAML manifest file to disk; it contains all of the required fields for the extension with example values. Later in the development lifecycle, the developer of a shinytableau extension can use values relevant to the project.

yaml_skeleton(filename = "manifest.yml")

Arguments

filename
A filename for the YAML manifest file. Can be combined with a relative path. The default, and recommended, filename is "manifest.yml".

Examples

# Generate a YAML manifest file in the
# root directory of the project
# yaml_skeleton()

# Generate the skeleton YAML manifest
# file (called `manifest.yml`) in the
# `app` subdirectory
# yaml_skeleton(filename = "app/manifest.yml")