Construct a reactive expression that reads Tableau data

This function is used to read data from Tableau. Because of the many levels of indirection involved in actually physically reading data from Tableau, using this function is significantly more involved than, say, a simple read.csv(). See the Details section for a more detailed introduction.

reactive_tableau_data(
  spec,
  options = list(),
  session = shiny::getDefaultReactiveDomain()
)

Arguments

spec
An argument that specifies what specific data should be retrieved. This can be specified in a number of ways:

The name of a setting, that was set using a value returned from choose_data(). This is the most common scenario for server.

The object returned from choose_data() can be passed in directly. This is likely the approach you should take if you want to retrieve data in config_server based on unsaved config changes (e.g. to give the user a live preview of what their choose_data choices would yield).

You can directly create a spec object using one of the helper functions spec_summary(), spec_underlying(), or spec_logical(). For cases where the data is not selected based on choose_data() at all, but programmatically determined or hardcoded. (This should not be common.)

options
A named list of options:

ignoreAliases - Do not use aliases specified in the data source in Tableau. Default is FALSE.

ignoreSelection - If FALSE (the default), only return data for the currently selected marks. Does not apply for datasource tables, only summary and underlying.

includeAllColumns - Return all the columns for the table. Default is FALSE. Does not apply for datasource and summary tables, only underlying.

maxRows - The maximum number of rows to return. Tableau will not, under any circumstances, return more than 10,000 rows for datasource and underlying tables. This option is ignored for summary tables.

columnsToInclude - Character vector of columns that should be included; leaving this option unspecified means all columns should be returned. Does not apply for summary and underlying, only datasource.

session
The Shiny session object. (You should probably just use the default.)

Details

There are two complicating factors when reading data from Tableau; the first is how to tell shinytableau what specific data table you want to access, and the second is actually accessing the data from R.

Specifying a data table

As of today, Tableau's Extension API requires you to access any data tables via one of the worksheets on the same dashboard.

Examples

data_spec_x <- choose_data("x", iv = iv)

#> Error in callModule(module, id, session = session): session must be a ShinySession or session_proxy object.

data_x <- reactive_tableau_data(data_spec_x)

#> Error in unwrap_session(session): shinytableau session not initialized. Are you trying to use tableau functions outside of a shinytableau extension?