Create and use a skeleton YAML manifest file

Tableau extensions need to provide a manifest (.trex) file that Tableau users can load into their dashboards. This file contains metadata such as the ID and name of the extension, the name and email of the extension author, logo image, etc. With shinytableau, you as the extension author do not create this file directly; instead, you provide a simpler manifest.yml file.

yaml_skeleton(filename = "manifest.yml")

tableau_manifest_from_yaml(path = ".")

Arguments

filename
The (relative or absolute) file path where the YAML manifest file should be written. For an app.R based app, the recommended file path is manifest.yml in the same directory as app.R; for a package-based app, put the file in the package's inst directory, and call tableau_manifest_from_yaml() with system.file().

path
The (relative or absolute) file or directory path from which to load an existing YAML manifest file. If the path is a directory, the filename manifest.yml will be added to the end of the path.

Value

yaml_skeleton returns the filename, invisibly. tableau_manifest_from_yaml returns a manifest object that is suitable for passing to tableau_extension().

Details

Call yaml_skeleton() from the R console to create a manifest.yml file.

Call tableau_manifest_from_yaml() from your app.R to load it, and pass the resulting object to tableau_extension().

See also

Examples

if (FALSE) {
# Generate a YAML manifest file in the
# root directory of the project
yaml_skeleton()

# Generate the skeleton YAML manifest
# file (called `manifest.yml`) in the
# `app` subdirectory
yaml_skeleton(filename = "app/manifest.yml")
}